+++ packages/pay-ark/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./ark-wallet";
+export * from "./interfaces";
 export * from "./network";
 export * from "./transaction";
-export * from "./ark-wallet";
