+++ packages/pay-commands/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./command";
 export * from "./commands";
+export * from "./interfaces";
