+++ packages/pay-currency/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./currency";
+export * from "./interfaces";
