+++ packages/pay-messenger/src/index.ts
@@ -1,2 +1,3 @@
-export * from "./messenger";
+export * from "./interfaces";
 export * from "./messages";
+export * from "./messenger";
