+++ packages/pay-messenger/src/interfaces.ts
@@ -1,0 +1,6 @@
+export interface Reply {
+    directMessageSender?: string;
+    directMessageReceiver?: string;
+    directMessageMerchant?: string;
+    replyComment?: string;
+}
