+++ packages/pay-reddit/src/index.ts
@@ -1,1 +1,2 @@
+export * from "./interfaces";
 export * from "./reddit-listener";
