+++ packages/pay-reddit/src/interfaces.ts
@@ -1,0 +1,24 @@
+export interface RedditConfig {
+    admin: string;
+    userAgent: string;
+    clientId: string;
+    clientSecret: string;
+    username: string;
+    password: string;
+    requestDelay: number;
+    continueAfterRatelimitError: boolean;
+    networks: string[];
+}
+
+export interface RedditMessage {
+    author: Author;
+    name: string;
+    body: string;
+    was_comment: boolean;
+    id: string;
+    parent_id?: string;
+}
+
+export interface Author {
+    name;
+}
