+++ packages/pay-storage/src/interfaces.ts
@@ -1,0 +1,5 @@
+export interface Wallet {
+    address: string;
+    encryptedSeed: string;
+    seed?: string;
+}
