+++ packages/pay-user/src/interfaces.ts
@@ -1,0 +1,4 @@
+export interface Username {
+    username: string;
+    platform: string;
+}
